adiciona e simplifica verificação de usuário logado

+++ src/main/java/br/com/cumbuca/service/usuario/UsuarioServiceImpl.java
@@ -1,6 +1,7 @@
 package br.com.cumbuca.service.usuario;
 
 import br.com.cumbuca.dto.usuario.UsuarioRequestDTO;
+import br.com.cumbuca.exception.CumbucaException;
 import br.com.cumbuca.model.Usuario;
 import br.com.cumbuca.repository.UsuarioRepository;
 import org.modelmapper.ModelMapper;
@@ -21,7 +22,7 @@ public class UsuarioServiceImpl implements UsuarioService {
     private final ModelMapper modelMapper;
 
     public UsuarioServiceImpl(UsuarioRepository usuarioRepository, PasswordEncoder passwordEncoder,
-            ModelMapper modelMapper) {
+                              ModelMapper modelMapper) {
         this.usuarioRepository = usuarioRepository;
         this.passwordEncoder = passwordEncoder;
         this.modelMapper = modelMapper;
@@ -30,7 +31,7 @@ public class UsuarioServiceImpl implements UsuarioService {
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
         return usuarioRepository.findByUsernameOrEmail(username, username)
-                .orElseThrow(() -> new UsernameNotFoundException("O usuário não foi encontrado."));
+                .orElseThrow(() -> new UsernameNotFoundException("Usuário não encontrado."));
     }
 
     @Override
@@ -45,7 +46,7 @@ public class UsuarioServiceImpl implements UsuarioService {
             try {
                 usuario.setFoto(usuarioRequestDTO.getFoto().getBytes());
             } catch (IOException e) {
-                throw new RuntimeException(e);
+                throw new CumbucaException("Erro ao processar a foto.");
             }
         }
 
@@ -55,14 +56,16 @@ public class UsuarioServiceImpl implements UsuarioService {
 
     @Override
     public Usuario atualizar(Long id, UsuarioRequestDTO usuarioRequestDTO) {
-        Usuario usuario = usuarioRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
+        final Usuario usuario = getUsuarioLogado();
+        if (!id.equals(usuario.getId())) {
+            throw new CumbucaException("Usuário não tem permissão para realizar esta ação.");
+        }
 
         if (usuarioRequestDTO.getFoto() != null && !usuarioRequestDTO.getFoto().isEmpty()) {
             try {
                 usuario.setFoto(usuarioRequestDTO.getFoto().getBytes());
             } catch (IOException e) {
-                throw new RuntimeException("Erro ao processar a foto.");
+                throw new CumbucaException("Erro ao processar a foto.");
             }
         }
 
@@ -71,7 +74,10 @@ public class UsuarioServiceImpl implements UsuarioService {
 
     @Override
     public void remover(Long id) {
-        Usuario usuario = usuarioRepository.findById(id).orElseThrow(() -> new CumbucaException("Usuário não encontrado"));
+        final Usuario usuario = getUsuarioLogado();
+        if (!id.equals(usuario.getId())) {
+            throw new CumbucaException("Usuário não tem permissão para realizar esta ação.");
+        }
         usuarioRepository.delete(usuario);
     }
 
@@ -85,6 +91,6 @@ public class UsuarioServiceImpl implements UsuarioService {
         }
 
         return usuarioRepository.findByUsernameOrEmail(login, login)
-                .orElseThrow(() -> new UsernameNotFoundException("O usuário não foi encontrado."));
+                .orElseThrow(() -> new UsernameNotFoundException("Usuário não encontrado."));
     }
 }
